Rename all composable files
