+++ frontend/src/components/UI/Tables/AuthorizationTable.vue
@@ -145,7 +145,7 @@
 import { ref, computed } from "vue";
 import { useStore } from "vuex";
 import { useQuasar } from "quasar";
-import useAuthorizationsTabsData from "../../../composables/TabPanels/useAuthorizationsTabs";
+import useAuthorizationsTabsData from "../../../composables/TabPanels/useAuthorizationsTabsData";
 import Modal from "../Dialogs/Modal.vue";
 import globalAvatar from "../../../assets/profil.png";
