+++ frontend/src/components/UI/Tables/EnvirTypeTable.vue
@@ -170,7 +170,7 @@ import { ref, computed } from "vue";
 import { useStore } from "vuex";
 import { useQuasar } from "quasar";
 import Modal from "../Dialogs/Modal.vue";
-import useFileData from "../../../composables/Forms/useFile";
+import useFileData from "../../../composables/Forms/userFileData";
 import FileUploader from "../Form/FileUploader.vue";
 import globalAvatar from "../../../assets/profil.png";
