+++ frontend/src/composables/TabPanels/useAuthorizationsTabsData.js
