+++ frontend/src/composables/TabPanels/useProductsTabData.js
